Hackathon/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.33.0
